Refactored servertools classes.
