+++ supriya/tools/servertools/BlockAllocator.py
@@ -64,11 +64,13 @@ class BlockAllocator(SupriyaObject):
         ):
         from supriya.tools import servertools
         from supriya.tools import timetools
-        self._free_heap = timetools.TimespanCollection()
+        self._free_heap = timetools.TimespanCollection(
+            accelerated=True)
         self._heap_maximum = heap_maximum
         self._heap_minimum = heap_minimum
         self._lock = threading.Lock()
-        self._used_heap = timetools.TimespanCollection()
+        self._used_heap = timetools.TimespanCollection(
+            accelerated=True)
         free_block = servertools.Block(
             start_offset=heap_minimum,
             stop_offset=heap_maximum,
@@ -151,7 +153,10 @@ class BlockAllocator(SupriyaObject):
         block_id = int(block_id)
         with self._lock:
             cursor = self._used_heap.get_simultaneity_at(block_id)
-            blocks = cursor.start_timespans + cursor.overlap_timespans
+            blocks = sorted(
+                set(cursor.start_timespans) or
+                set(cursor.overlap_timespans)
+                )
             assert len(blocks) == 1
             used_block = blocks[0]
             self._used_heap.remove(used_block)
