+++ supriya/tools/servertools/Group.py
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 from __future__ import print_function
 import collections
-from supriya.tools.datastructuretools.TreeContainer import TreeContainer
+from supriya.tools.systemtools.TreeContainer import TreeContainer
 from supriya.tools.servertools.Node import Node
